Simple proof of liabilities
Quickly made, but should work well.

Users identified by appending their email to their user id and sha256
hashing.
This should prevent a bruteforce attack revealing the emails of users.

Couple this with a link the the cold storage address, and a signed
message would make it very easy to check solvency.

Obviously in production you would create a cron job that makes a HTML
version every 30 minutes or so. Easily adapted to support multiple
currencies.

Definitely not the most efficient way, but functional.