Fix for authentication issue
Proposed fix for the failing authentication of the API.

After much checking, determined that PHPs JSON parser is largely broken, and
considers it to be an array. If this is the case (it is) it adjusts by taking the first
(and only) element of the array, and setting the $api_signature1 to the value
before continuing on its general checks of the signature.